# perf: Optimize IsNotNullExpr