--- a/datafusion/physical-expr/src/expressions/is_null.rs
+++ b/datafusion/physical-expr/src/expressions/is_null.rs
@@ -120,13 +120,8 @@ pub(crate) fn compute_is_null(array: ArrayRef) -> Result<BooleanArray> {
 /// workaround <https://github.com/apache/arrow-rs/issues/6017>,
 /// this can be replaced with a direct call to `arrow::compute::is_not_null` once it's fixed.
 pub(crate) fn compute_is_not_null(array: ArrayRef) -> Result<BooleanArray> {
-    if let Some(union_array) = array.as_any().downcast_ref::<UnionArray>() {
-        let is_null = if let Some(offsets) = union_array.offsets() {
-            dense_union_is_null(union_array, offsets)?
-        } else {
-            sparse_union_is_null(union_array)?
-        };
-        compute::not(&is_null).map_err(Into::into)
+    if array.as_any().is::<UnionArray>() {
+        compute::not(&compute_is_null(array)?).map_err(Into::into)
     } else {
         compute::is_not_null(array.as_ref()).map_err(Into::into)
     }